+++ engine/baml-runtime/src/tracing/mod.rs
@@ -199,12 +199,11 @@ impl BamlTracer {
         let trace_stats = TraceStats::default();
 
         let tracer = BamlTracer {
-            // tracer: if options.enabled() {
-            //     Some(TracerImpl::new(&options, 20, trace_stats.clone()))
-            // } else {
-            //     None
-            // },
-            tracer: None,
+            tracer: if options.enabled() {
+                Some(TracerImpl::new(&options, 20, trace_stats.clone()))
+            } else {
+                None
+            },
             options,
             trace_stats,
         };
