+++ integ-tests/python/tests/test_pydantic.py
@@ -1,5 +1,4 @@
 import baml_py
-import inspect
 import pydantic
 import pytest
 
@@ -79,23 +78,3 @@ def test_model_dump():
             "media_type": "image/png",
         }
     }
-
-
-def test_inspect():
-    """Assert that baml_py is compatible with the inspect module.
-
-    This is a regression test for a bug where `inspect.stack()` would implode if the
-    pyo3 code called `PyModule::from_code` without specifying the `file_name` arg (i.e.
-    without specifying the source file metadata for the inline Python snippet).
-    """
-
-    class LoremIpsum(pydantic.BaseModel):  # pyright: ignore[reportUnusedClass]
-        """Defining this Pydantic model alone is sufficient to trigger the bug."""
-
-        my_image: baml_py.Image
-        my_audio: baml_py.Audio
-
-    try:
-        inspect.stack()
-    except Exception as e:
-        pytest.fail(f"inspect.stack() raised an unexpected exception: {e}")
